fix: redirect if the marker code changes
When updating a marker location there's the possibility that the marker
overlaps with another location. In that case, the location service
creates a new marker and updates the marker and location codes. This
also includes a style update that warns the user and indicates autosave
is disabled for map updates